Reorganizing project and demo - WIP
