+++ demo/__init__.py
@@ -1,0 +1,4 @@
+import os
+import sys
+
+sys.path.append(os.path.abspath(os.getcwd() + os.sep + "src"))
