+++ src/throttle/__init__.py
@@ -1,4 +1,4 @@
-from . import registries
+from . import predefined
 from .settings import DefaultSettings, Settings
 
-__all__ = ["Settings", "DefaultSettings", "registries"]
+__all__ = ["Settings", "DefaultSettings", "predefined"]
